+++ src/text/reexports.rs
@@ -17,7 +17,16 @@ impl_cdef!["" => &str];
 #[cfg(feature = "alloc")]
 impl_cdef![Self::new() => String];
 
-/* alloc, std */
+/* core, alloc */
+
+reexport! { rust: core,
+    doc: "Constructs parameters for the other string-formatting macros.",
+    format_args
+}
+reexport! { rust: alloc,
+    doc: "Creates a String using interpolation of runtime expressions.",
+    format
+}
 
 reexport! { rust: alloc::string,
     doc: "A UTF-8–encoded, growable string.",